13: Move the Reflection helper class into the context package and only allow the app context to interact with it
